+++ test/dummy/test/integration/root_switch_dropdown_test.rb
@@ -36,7 +36,9 @@ class RootSwitchDropdownTest < ActionDispatch::IntegrationTest
 
     assert_response :success
     assert_includes response.body, workspace.name
-    assert_select "body[data-recording-studio-default-layout='true']", count: 1
+    assert_includes response.body, "data-controller=\"flat-pack--sidebar-layout\""
+    assert_includes response.body, "recording-studio-root-switchable--root-switch-dropdown"
+    refute_includes response.body, "data-recording-studio-default-layout"
   end
 
   test "root switch page renders with the host default layout" do